Changed doubles to floats for feature vector values. Chagned data to be sourced directly from URLs. TinyImageFloatFV is now deprecated, replaced by NormalisableFloatFV which doesn't change the values in place when it's normalising function is called. Began fleshing out Run1 and Run2 skeletons.

+++ src/main/java/uk/ac/soton/ecs/coursework3/FeatureVectorClassPair.java
@@ -4,10 +4,10 @@ package uk.ac.soton.ecs.coursework3;
  * Simple class containing a feature vector and the name of the class it belongs to.
  */
 public class FeatureVectorClassPair {
-    public double[] featureVector;
+    public float[] featureVector;
     public String vectorClass;
 
-    public FeatureVectorClassPair(double[] featureVector, String vectorClass) {
+    public FeatureVectorClassPair(float[] featureVector, String vectorClass) {
         this.featureVector = featureVector;
         this.vectorClass = vectorClass;
     }
